refactor: Seperate account service into account and auth
