+++ src/main/java/com/cherrydev/cherrymarketbe/account/service/AccountServiceImpl.java
@@ -6,13 +6,6 @@ import com.cherrydev.cherrymarketbe.account.repository.AccountMapper;
 import com.cherrydev.cherrymarketbe.auth.dto.oauth.OAuthAccountInfoDto;
 import com.cherrydev.cherrymarketbe.common.exception.AuthException;
 import com.cherrydev.cherrymarketbe.common.exception.DuplicatedException;
-import com.cherrydev.cherrymarketbe.common.exception.NotFoundException;
-import com.cherrydev.cherrymarketbe.common.jwt.JwtProvider;
-import com.cherrydev.cherrymarketbe.common.jwt.dto.JwtReissueResponseDto;
-import com.cherrydev.cherrymarketbe.common.jwt.dto.JwtRequestDto;
-import com.cherrydev.cherrymarketbe.common.jwt.dto.JwtResponseDto;
-import com.cherrydev.cherrymarketbe.common.service.RedisService;
-import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -20,9 +13,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.cherrydev.cherrymarketbe.account.dto.AccountInfoDto;
-import com.cherrydev.cherrymarketbe.auth.dto.SignInRequestDto;
-import com.cherrydev.cherrymarketbe.auth.dto.SignInResponseDto;
-import com.cherrydev.cherrymarketbe.auth.dto.SignUpRequestDto;
+import com.cherrydev.cherrymarketbe.auth.dto.*;
 import com.cherrydev.cherrymarketbe.account.entity.Account;
 
 import java.util.Arrays;
@@ -30,10 +21,8 @@ import java.util.Arrays;
 import static com.cherrydev.cherrymarketbe.account.enums.RegisterType.*;
 import static com.cherrydev.cherrymarketbe.account.enums.UserRole.ROLE_CUSTOMER;
 import static com.cherrydev.cherrymarketbe.account.enums.UserStatus.*;
-import static com.cherrydev.cherrymarketbe.common.constant.AuthConstant.*;
 import static com.cherrydev.cherrymarketbe.common.exception.enums.ExceptionStatus.*;
 import static com.cherrydev.cherrymarketbe.common.utils.CodeGenerator.generateRandomCode;
-import static org.springframework.beans.propertyeditors.CustomBooleanEditor.VALUE_TRUE;
 
 
 @Service
@@ -44,8 +33,6 @@ public class AccountServiceImpl implements AccountService {
 
     private final AccountMapper accountRepository;
     private final PasswordEncoder passwordEncoder;
-    private final JwtProvider jwtProvider;
-    private final RedisService redisService;
 
     @Override
     @Transactional
@@ -88,11 +75,11 @@ public class AccountServiceImpl implements AccountService {
     @Override
     @Transactional(readOnly = true)
     public ResponseEntity<AccountInfoDto> getAccountInfo(final AccountDetails accountDetails) {
-        Account requestedAccount = accountDetails.getAccount();
+        Account account = accountDetails.getAccount();
         return ResponseEntity
                 .ok()
                 .body(
-                        new AccountInfoDto(requestedAccount)
+                        new AccountInfoDto(account)
                 );
     }
 
@@ -108,105 +95,22 @@ public class AccountServiceImpl implements AccountService {
         accountRepository.delete(accountDetails.getAccount());
     }
 
-    @Override
-    @Transactional
-    public ResponseEntity<SignInResponseDto> signIn(
-            final SignInRequestDto signInRequestDto
-    ) {
-        String email = signInRequestDto.getEmail();
-        String requestedPassword = signInRequestDto.getPassword();
-        Account account = findAccountByEmail(email);
-
-        checkUserStatusByEmail(account);
-        checkPasswordIsCorrect(requestedPassword, account);
-
-        final JwtResponseDto jwtResponseDto = jwtProvider.createJwtToken(email);
-        redisService.setDataExpire(email, jwtResponseDto.getRefreshToken(),
-                REFRESH_TOKEN_EXPIRE_TIME);
-
-        return ResponseEntity
-                .ok()
-                .header(AUTHORIZATION_HEADER, BEARER_PREFIX + jwtResponseDto.getAccessToken())
-                .body(
-                        SignInResponseDto.builder()
-                                .userName(account.getName())
-                                .userRole(account.getUserRole())
-                                .grantType(jwtResponseDto.getGrantType())
-                                .accessToken(jwtResponseDto.getAccessToken())
-                                .refreshToken(jwtResponseDto.getRefreshToken())
-                                .expiresIn(jwtResponseDto.getAccessTokenExpiresIn())
-                                .build()
-                );
-    }
-
-    /**
-     * 로그아웃 처리를 위해 사용자 인증 수단(토큰)을 검증하고, 무효화한다.
-     */
-    @Override
-    @Transactional
-    public void signOut(final JwtRequestDto jwtRequestDto) {
-        jwtProvider.validateToken(jwtRequestDto.getAccessToken());
-
-
-        Claims claims = jwtProvider.getInfoFromToken(jwtRequestDto.getAccessToken());
-        String email = claims.getSubject();
-
-        invalidateToken(email, jwtRequestDto.getAccessToken());
-    }
-
-    @Override
-    @Transactional
-    public ResponseEntity<JwtReissueResponseDto> reissue(final JwtRequestDto jwtRequestDto) {
-
-        validateRefreshToken(jwtRequestDto);
-        validateRefreshTokenOwnership(jwtRequestDto);
-
-        String email = jwtProvider.getInfoFromToken(jwtRequestDto.getAccessToken()).getSubject();
-
-        JwtResponseDto jwtResponseDto = jwtProvider.createJwtToken(email);
-
-        return ResponseEntity
-                .ok()
-                .header(AUTHORIZATION_HEADER, BEARER_PREFIX + jwtResponseDto.getAccessToken())
-                .body(
-                        JwtReissueResponseDto.builder()
-                                .accessToken(jwtResponseDto.getAccessToken())
-                                .refreshToken(jwtRequestDto.getRefreshToken())
-                                .accessTokenExpiresIn(jwtResponseDto.getAccessTokenExpiresIn())
-                                .build()
-                );
-    }
-
     // =============== PRIVATE METHODS =============== //
 
-    private void checkPasswordIsCorrect(String requestedPassword, Account account) {
-        if (!passwordEncoder.matches(requestedPassword, account.getPassword())) {
-            throw new AuthException(INVALID_ID_OR_PW);
-        }
-    }
-
-    private void checkUserStatusByEmail(Account account) {
-        if (account.getUserStatus().equals(DELETED)) {
-            throw new NotFoundException(DELETED_ACCOUNT);
-        }
-        if (account.getUserStatus().equals(RESTRICTED)) {
-            throw new AuthException(RESTRICTED_ACCOUNT);
-        }
-    }
-
-    private Account findAccountByEmail(final String email) {
-        return accountRepository.findByEmail(email)
-                .orElseThrow(() -> new NotFoundException(NOT_FOUND_ACCOUNT));
-    }
-
+    /**
+     * 회원 가입 시 이메일이 중복되는지 확인한다.
+     */
     private void checkEmailIsDuplicated(String email) {
         if (accountRepository.existByEmail(email)) {
             throw new DuplicatedException(CONFLICT_ACCOUNT);
         }
     }
 
+    /**
+     * 회원 가입 시 이름에 금지어가 포함되어 있는지 확인한다.
+     * @see ForbiddenUserName 금지어 목록
+     */
     private void checkUsernameIsProhibited(String username) {
-        // 금지 이름을 contain하는지 검사
         boolean isProhibited = Arrays.stream(ForbiddenUserName.values())
                 .anyMatch(forbiddenUserName -> username.contains(forbiddenUserName.getName()));
 
@@ -215,31 +119,4 @@ public class AccountServiceImpl implements AccountService {
         }
     }
 
-    /**
-     * 토큰의 유효성을 검증한다.
-     */
-    private void validateRefreshToken(JwtRequestDto jwtRequestDto) {
-        jwtProvider.validateToken(jwtRequestDto.getRefreshToken());
-    }
-
-    /**
-     * 요청자와 토큰의 소유자 정보가 일치하는지 검증한다.
-     */
-    private void validateRefreshTokenOwnership(JwtRequestDto jwtRequestDto) {
-        String email = jwtProvider.getInfoFromToken(jwtRequestDto.getAccessToken()).getSubject();
-        String validRefreshToken = redisService.getData(email);
-        if (!jwtRequestDto.getRefreshToken().equals(validRefreshToken)) {
-            throw new AuthException(EXPIRED_REFRESH_TOKEN);
-        }
-    }
-
-    /**
-     * REDIS에서 사용자의 정보를 삭제하고, 토큰을 BLACK_LIST에 추가해 토큰을 무효화한다.
-     */
-    private void invalidateToken(String email, String accessToken) {
-        redisService.deleteData(email);
-        redisService.setDataExpire(
-                BLACK_LIST_KEY_PREFIX + accessToken,
-                VALUE_TRUE, ACCESS_TOKEN_EXPIRE_TIME / 1000L);
-    }
 }
